refactor(response): request/response handling to authless Response
